Added dirty file checks for tags and dates This patch brings the previous together, adding code needed to check for dirty src files, removed src files, or changed tags and dates.
If any of these conditions are true, the group needs to be rebuilt.
But for the vast majority of cases when just editing a single update
to the blog, none of these files need to be changed.

With all these patches, and using memcached, blog build times on my
system, containing about 75 entries, about 40 tags, and dates from
2005 to 2010, a cold build takes 2 minutes 4 seconds.  A cached
build takes 9 seconds.
---
 bin/chronicle |  108 +++++++++++++++++++++++++++++++++++++++++++++++++++++---
 1 files changed, 102 insertions(+), 6 deletions(-)