Added new cross reference feature New header key "xrefs:" added to the entry files.  This is an optional comma separated list of base entry filenames (no directories) of old posts that the current post is linking to.
The actual href links still need to be added manually to the current
post.

Each linked post is updated with a list of links at the bottom,
pointing to other posts that link back to it.

This creates an automatic forward-looking chain of links to help
readers follow logically related blog posts.

Bug note: memcache behaviour works as expected when adding xrefs, but
not when removing them.  To remove a xref, a non-cached build is
required.
---
 bin/chronicle |  148 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 1 files changed, 148 insertions(+), 0 deletions(-)